Refactor chat settings to Hilos::$setting
- Add catalog-backed typed readers for string, int, float, and bool settings.
- Introduce a SettingException hierarchy for settings access failures.
- Wire project settings through the Hilos facade extension point.
- Update chat agents, upload limits, settings UI, and integration coverage.
- Document framework-development rules and add the matching Codex skill.

+++ demo/chat/backend/Database/Migration/Seed/003_chat_settings.sql
@@ -7,11 +7,11 @@ INSERT INTO `hilos_setting` (`key`, `type`, `value`) VALUES
 ('chat_bot_model', 'string', 'qwen2.5:3b'),
 ('chat_bot_url', 'string', ''),
 ('chat_bot_provider', 'string', 'local'),
-('chat_bot_timeout_sec', 'integer', '90'),
+('chat_bot_timeout_sec', 'float', '90'),
 ('chat_bot_language', 'string', 'ru'),
 ('chat_moderation_model', 'string', 'qwen2.5:0.5b'),
 ('chat_moderation_url', 'string', ''),
 ('chat_moderation_provider', 'string', 'local'),
-('chat_moderation_timeout_sec', 'integer', '90'),
+('chat_moderation_timeout_sec', 'float', '90'),
 ('orphan_test_demo', 'string', 'leftover_from_migration')
 ON DUPLICATE KEY UPDATE `type` = VALUES(`type`), `value` = VALUES(`value`);
